requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Minimal Ultralytics runtime set for this inference-only API.
pillow==10.4.0
numpy==1.26.4
matplotlib==3.10.8
opencv-python-headless==4.10.0.84
psutil==7.2.2
pyyaml==6.0.3
requests==2.32.5
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3

# Install this with --no-deps in Docker to avoid training/analytics extras
# like pandas, seaborn, scipy, py-cpuinfo, and ultralytics-thop.
ultralytics==8.3.0
